social_media_agent_platform/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
alembic==1.12.1
psycopg2-binary==2.9.9
firebase-admin==6.2.0
langchain==0.0.335
openai==1.3.0
python-twitter==3.5
facebook-sdk==3.1.0
python-linkedin-v2==0.3.3
google-cloud-firestore==2.11.1
google-cloud-secret-manager==2.16.4
pytest==7.4.2
httpx==0.25.0
